requirements.txt
torch
torchvision
kaggle
python-dotenv
Pillow
opencv-python
facenet-pytorch
matplotlib